# switch from conda to uv