--- a/images/plbase/Dockerfile
+++ b/images/plbase/Dockerfile
@@ -1,5 +1,7 @@
 FROM amazonlinux:2023
 
+ENV PATH="/.venv/bin:$PATH"
+
 # plbase-install.sh: script to install all needed packages and dependencies
 # python-requirements.txt: list of python packages to be installed by pip
 COPY plbase-install.sh python-requirements.txt /