- From local testing, build times are ~25% faster. (240seconds to 190seconds). I attribute a lot of this to uv having a more performant pip install than the default
- Docker image was ~13% smaller (2742682177 bytes -> 2385065557 bytes) via `docker image inspect plbase:latest --format='{{.Size}}'`

- Installs a version of Python, and sets up a venv to use with it.