@eliotwrobson 2024-12-19
Haven't looked too deeply into this, but if things are going faster, I think this is probably the right direction to go since we aren't really using the features that make conda worthwhile. Might be worth folding into https://github.com/PrairieLearn/PrairieLearn/pull/10984 (but separate from https://github.com/PrairieLearn/PrairieLearn/pull/11033 which should be merged first)