@reteps 2024-12-19
```
Error page {"id":"V5BMVEUEOIS2","msg":"pool is draining and cannot accept work","referrer":null,"response_id":"ab4dc258-e8f1-440d-b27d-aa64719165bd","stack":"Error: pool is draining and cannot accept work\n    at Pool.acquire (/PrairieLearn/node_modules/generic-pool/lib/Pool.js:447:9)\n    at withCodeCaller (/PrairieLearn/apps/prairielearn/src/lib/code-caller/index.ts:160:33)\n    at file:///PrairieLearn/apps/prairielearn/src/question-servers/freeform.js:1088:18\n    at async <anonymous> (/PrairieLearn/packages/opentelemetry/src/tracing.ts:11:24)\n    at async makeVariant (/PrairieLearn/apps/prairielearn/src/lib/question-variant.ts:91:57)\n    at async makeAndInsertVariant (/PrairieLearn/apps/prairielearn/src/lib/question-variant.ts:225:50)\n    at async ensureVariant (/PrairieLearn/apps/prairielearn/src/lib/question-variant.ts:369:10)\n    at async getAndRenderVariant (/PrairieLearn/apps/prairielearn/src/lib/question-render.ts:429:22)\n    at async <anonymous
```

```
 @prairielearn/prairielearn:test: error: load.endJob(): python_worker_idle no such id: 1c0d5fd2-8c94-4fc7-8c86-364f2fde0dd1
@prairielearn/prairielearn:test: error: load.endJob(): python_worker_idle no such id: a256cde9-7df9-46a1-8fc8-c5ac745cbdd6
```

```
 @prairielearn/prairielearn:test:      Error: Timeout of 10000ms exceeded. For async tests and hooks, ensure "done()" is called; if returning a Promise, ensure it resolves. (/PrairieLearn/apps/prairielearn/src/tests/groupScoreAndSync.test.ts)
@prairielearn/prairielearn:test:       at listOnTimeout (node:internal/timers:581:17)
@prairielearn/prairielearn:test:       at process.processTimers (node:internal/timers:519:7)
```
(obvious test regression in code-caller)