@eliotwrobson 2025-01-03
A little preemptive, but worth thinking about whether we want to use lock files with uv and whether this can save some time on installation: https://docs.astral.sh/uv/pip/compile/