@echuber2 2025-01-14
Looks like the earlier version that just installed "3.10" also grabbed the cpython version by default, judging from one of the earlier jobs that had timed out:

```
136.2 + uv python install --default --preview 3.10
137.6 Installed Python 3.10.16 in 1.36s
137.6  + cpython-3.10.16-linux-x86_64-gnu (python, python3, python3.10)
```

It may be worth writing a switch to flip between the "most optimized" ARM or "most optimized" x86 build available depending on the target platform, but I'm not sure how best to detect that with the GitHub runner setup. And, it might be ugly to write in the latest minor version that way.