@echuber2 2025-01-14
Since Amazon Linux 2023 doesn't have 3.10 as a package (apparently), it might also be worth trying uv on top of the conda/mamba build of the runtime.