@reteps 2025-01-14
https://github.com/astral-sh/uv/blob/main/crates/uv-python/src/managed.rs#L251 -- this looks like why.