@echuber2 2025-01-14
I found that moving `export UV_PYTHON_PREFERENCE=only-managed` after the venv is created helps to get around the issue, or it could simply be removed. (I was about to push another experiment if that's okay.)