@reteps 2025-01-14
I created an issue explaining this behavior. The optimized builds were only added in December 2024, so this may just be an oversight.