@echuber2 2025-01-14
Welp, this is still timing out. I wonder if the Python version is really the issue. Maybe it will help to update based on master, or try layering `uv pip install` on top of the previous miniforge build.