@echuber2 2025-01-14
It may aide debugging to split out all of the uv stuff to a separate `RUN` script after the `dnf` and `npm` parts, at least temporarily. This makes rebuilds much faster locally.