@reteps 2025-01-14
> Welp, this is still timing out. I wonder if the Python version is really the issue. Maybe it will help to update based on master, or try layering `uv pip install` on top of the previous miniforge build.

I'm not super excited about having to keep conda around. I'll probably do some more digging to see if there is another reason why the python is slower (should it really be that much slower????)