@echuber2 2025-01-14
It might be worth rebasing on master first, and if the timeout still happens, temporarily use miniforge just to see if it too times out. Maybe another change somewhere is causing that.