@eliotwrobson 2025-01-26
@reteps looks like the fix for the issue you reported got merged.