@echuber2 2025-02-11
I added the entry for Linux aarch64 which is what I suppose building on macOS actually would require, but I left the Darwin target also in case somebody wants to try running the install script natively on Mac too.

On the speed issue, it looks like the x86_64 v3 version _does_ have PGO and LTO optimizations (after inspecting `python3 -m sysconfig`), so if it's still slow, there must be something these builds are doing differently from the conda distro.