@echuber2 2025-02-11
...it's also possible the PGO and LTO optimizations might somehow be making it slower, so it could be worth trying to manually download and install the Astral build that has no optimizations. It doesn't look like the "install" command can grab this one automatically, after studying the manifest their tool uses. However they do release tarballs of the unoptimized builds on their GitHub. Their documentation describes the `noopt` version as "A regular optimized build without PGO or LTO" so I'm wondering if the _regular_ optimizations are sufficient.