@eliotwrobson 2025-02-11
@echuber2 feel free to play around, I think we got pretty stuck and it would be much appreciated if you could resolve this 🙏