@echuber2 2025-02-11
This worked by disabling parallel tests in mocha. I suspect something more complicated is going on, or it might be the old bug here (PR rejected because I didn't have a good test case): https://github.com/mochajs/mocha/pull/4959 Not sure what to do.