@echuber2 2025-02-11
Someone else [suggested there might be a conflict with sinon](https://stackoverflow.com/questions/44149096/for-async-tests-and-hooks-ensure-done-is-called-if-returning-a-promise-en#comment120189727_66889727) but I don't want to grasp at straws.