@reteps 2025-05-08
on mac:
without UV:
![CleanShot 2025-05-08 at 15 50 23@2x](https://github.com/user-attachments/assets/837ea2e3-4c15-4309-89f0-e34538dd2d20)

with UV:

```
~/repos/PrairieLearn/apps/prairielearn stengerp/render-test *20 ❯ hyperfine 'yarn run mocha src/tests/render.test.ts && killall python3.10'                                                       PrairieLearn
Benchmark 1: yarn run mocha src/tests/render.test.ts && killall python3.10
  Time (mean ± σ):     41.706 s ±  0.402 s    [User: 46.682 s, System: 8.244 s]
  Range (min … max):   41.385 s … 42.587 s    10 runs
  ```