@nwalters512 2025-11-20
I did some benchmarking with the different images and Python versions. TL;DR I think we can drop the architecture-specific Python version selection.

---

`demo/randomPlot` in example course, tested 100 times:

with v3 build:
```
Generate/prepare: 300 tests, min 341ms, avg 378.07ms, max 738ms
Render:           300 tests, min 174ms, avg 192.58ms, max 476ms
Parse/grade:      300 tests, min 353ms, avg 522.27ms, max 1229ms
```

with default build:
```
Generate/prepare: 300 tests, min 353ms, avg 393.45ms, max 582ms
Render:           300 tests, min 174ms, avg 202.39ms, max 255ms
Parse/grade:      300 tests, min 360ms, avg 543.42ms, max 788ms
```

current master image:
```
Generate/prepare: 300 tests, min 417ms, avg 435.22ms, max 524ms
Render:           300 tests, min 212ms, avg 224.21ms, max 297ms
Parse/grade:      300 tests, min 433ms, avg 598.78ms, max 731ms
```

---

`addNumbers` in test course, tested 100 times:

with v3 build:
```
Generate/prepare: 300 tests, min 12ms, avg 13.47ms, max 35ms
Render:           300 tests, min 11ms, avg 13.14ms, max 22ms
Parse/grade:      300 tests, min 25ms, avg 35.01ms, max 141ms
```

with default branch:
```
Generate/prepare: 300 tests, min 12ms, avg 13.35ms, max 27ms
Render:           300 tests, min 12ms, avg 13.29ms, max 26ms
Parse/grade:      300 tests, min 25ms, avg 35.22ms, max 130ms
```

current master image:
```
Generate/prepare: 300 tests, min 12ms, avg 13.71ms, max 53ms
Render:           300 tests, min 12ms, avg 13.85ms, max 41ms
Parse/grade:      300 tests, min 25ms, avg 35.9ms, max 127ms
```